chain-api/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
PyJWT==2.9.0
pymongo==4.12.1
sqlparse==0.5.3
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
PyJWT==2.9.0
pymongo==4.12.1
python-dotenv==1.1.0
sqlparse==0.5.3
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.12.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
pyaes==1.6.1
PyJWT==2.9.0
pymongo==4.12.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
whitenoise==6.9.0
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
pyaes==1.6.1
PyJWT==2.9.0
pymongo==4.12.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.0
sqlparse==0.5.3
TgCrypto==1.2.5
typing_extensions==4.13.2
whitenoise==6.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
bitarray==3.4.1
certifi==2025.4.26
charset-normalizer==3.4.2
ckzg==2.1.1
cytoolz==1.0.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
frozenlist==1.6.0
gunicorn==23.0.0
hexbytes==1.3.1
idna==3.10
multidict==6.4.3
packaging==25.0
parsimonious==0.10.0
propcache==0.3.1
psycopg2-binary==2.9.10
py-solc-x==1.1.1
pyaes==1.6.1
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pymongo==4.12.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.0
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
semantic-version==2.10.0
sqlparse==0.5.3
TgCrypto==1.2.5
toolz==1.0.0
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
web3==7.11.1
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.0
